### 2022-08-11 #4670 @danielskatz
@cariyurek - I'm not sure exactly what the problem is here. You might want to try removing the blank lines in your paper's metadata section and more closely following the [JOSS example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).  Also, your paper needs a "Statement of Need" section.
