+++ paper.md
@@ -1,78 +1,63 @@
-# PyGRASP: A library to reconstruct Dynamic Contrast enhanced MR Images acquired with radial sampling
-
 ---
 title: 'PyGRASP: A library to reconstruct dynamic contrast enhanced MR images acquired with radial sampling'
-
 tags:
   - Python
   - dynamic contrast enhanced MRI
   - radial sampling
   - GRASP
-
 authors:
-- name: Cemre Ariyurek
+  - name: Cemre Ariyurek
     orcid: 0000-0002-1691-7097
     equal-contrib: true
     corresponding: true 
     affiliation: 1
-    
-- name: Aziz Ko\c{c}anao\u{g}ullar\i
+  - name: Aziz Ko\c{c}anao\u{g}ullar\i
     orcid: 0000-0002-4776-4206
     equal-contrib: true
     affiliation: 1
-    
-- name: Can Taylan Sari
-  orcid: 0000-0001-6140-3717
-  affiliation: 1
-  
-- name:  Serge Vasylechko
-  orcid: 0000-0002-5691-0607
-  affiliation: 1
-  
-- name:  Onur Afacan 
-  orcid: 0000-0003-2112-3205
-  affiliation: 1
-  
-- name:  Sila Kurugol 
-  orcid: 0000-0002-5081-4569
-  affiliation: 1
-  
- affiliations:
+  - name: Can Taylan Sari
+    orcid: 0000-0001-6140-3717
+    affiliation: 1      
+    - name:  Serge Vasylechko
+    orcid: 0000-0002-5691-0607
+    affiliation: 1
+    - name:  Onur Afacan 
+    orcid: 0000-0003-2112-3205
+    affiliation: 1
+    - name:  Sila Kurugol 
+    orcid: 0000-0002-5081-4569
+    affiliation: 1
+affiliations:
  - name: Quantitative Intelligent Imaging (QUIN) Lab, Boston Children's Hospital and Harvard Medical School
    index: 1
-   
 date: 10 August 2022
-
 bibliography: references.bib
-
 ---
 
-##Summary 
+# Summary
 
-In dynamic contrast-enhanced (DCE) magnetic resonance imaging (MRI), a dynamic series of image stacks are acquired capturing the passage of contrast agent through the tissue compartments.  Dynamic radial imaging with a golden angle stack of stars sampling trajectory has been increasingly used to acquire DCE-MRI data due to its advantages compared to dynamic cartesian imaging, such as being robust to breathing motion, and ability to achieve high temporal resolution imaging. In order to reconstruct a dynamic series of volumes at high temporal resolution (~3 seconds/volume), the data needs to be reconstructed from undersampled measurements. Different regularizers have been proposed to reduce the incoherent undersampling artifacts over the dynamic image series in radial acquisition, which are in the form of streaking artifacts [@Jung:2010][@Uecker:2010][@Poddar:2016][@Prieto:2010]. Golden-angle RAdial Sparse Parallel (GRASP) [@Feng:2014] reconstruction algorithm uses temporal regularization to reduce artifacts and can reconstruct a dynamic series of volumes at the desired temporal resolution.  In this software, we have developed an efficient open-source, purely Python-based GRASP reconstruction library called pyGRASP that allows researchers to access the source code for development, facilitating flexible deployment with readable code and no compilation. This library includes preprocessing by extracting information from raw data acquired in Siemens MRI scanners and coil sensitivity map estimation, coil compression and coil removal options for improving speed and image quality and image reconstruction using a compressed sensing approach based on GRASP technique, followed by post-processing. The proposed library has options to accelerate computations by parallel processing using either GPU or CPU resources. We also provide docker files for quick and easy deployment over any operating system or over a cloud.
- 
-##Statement of need
[…]
