+++ paper.md
@@ -49,8 +49,7 @@ pyGRASP has been used in [@AriyurekISMRMMoCo:2022], the coil selection method an
 
 # Algorithm
 
-pyGRASP consists of three modules. The first one reads the raw data acquired in Siemens MRI scanners, writes the k-space data into a file and calculates coil sensitivity maps based on Walsh method[@Zhang:2014] and the code is available in Github repository of ismrmrd-python-tools[@ismrmdpytools]. There are options for coil rejection and coil compression that can be enabled by the user. For coil rejection, the quality of information each coil carries is evaluated based on a mutual information based metric computed between a reference image (which is a combination of all data) and each coil image, and a set of coils below the threshold are removed, reducing streaking artifacts and improving the SNR of the reconstructed image [@KocanaogullariISMRMCoil:2022]. For coil compression, PCA is employed by compressing data from many channels into fewer virtual coils, with the goal of reducing the computational times. In the second module, GRASP reconstruction is computed, solving Eqn. [1]: 
-$$f(X) = \| FC x_t - k_t\|_2^2 + \lambda\| TV(X)\|_{1,1}$$ where $X = \lbrace x_t | t\in\lbrace 1,2,\cdots ,N \rbrace \rbrace$, $F$ is the nonuniform Fourier transform (NUFT), $C$ is the coil profile transform, $TV(.)$ is the total variation, $k_t$ and $x_t$ are the k-space data and image for a time frame t, respectively.  The first term in Eqn. [1] is the loss function for the data consistency and the second term is the total variation of reconstructed volumes in temporal direction used as a sparsifying transform followed by computation of its L1 norm for enforcing sparsity in time. Users can tune the weight of ($\lambda$) for the regularization term (second term) and choose the number of spokes per volume to adjust temporal resolution of reconstructed volumes. In pyGRASP, we enabled parallelization over dynamic volumes as observed in flowchart of the pyGRASP (\autoref{fig:fig1}).  GPU parallelization reduces the computation time to ~3.5 mins/slice using a batch size of 10 ( GPU’s memory was 24564MiB), whereas computation time was ~ 7.3 mins/slice when it was run on CPU with 32 cores. On the other hand, computation time of MATLAB based GRASP was very long with ~99 mins/slice for a dataset of 8 mins of data including 2600 spokes, 21 slices, 38 channels, and a matrix size of 448. 
[…]
