+++ references.bib
@@ -97,7 +97,7 @@
   title = {{TorchKbNufft}: A High-Level, Hardware-Agnostic Non-Uniform Fast {Fourier} Transform},
   booktitle = {ISMRM Workshop on Data Sampling \& Image Reconstruction},
   year = 2020,
-  note = {Source code available at https://github.com/mmuckley/torchkbnufft}.
+  note = {Source code available at https://github.com/mmuckley/torchkbnufft}
 }
 
 @misc{ismrmdpytools,
