### 2022-08-11 #4670 @danielskatz
@cariyurek - can you try removing the blank link in a bib entry (line 24) in your bib file, then try this command again?
