+++ references.bib
@@ -9,7 +9,6 @@
   month    =  apr,
   year     =  2016
 }
-
 @article{Prieto:2010,
   title    = "{3D} undersampled golden-radial phase encoding for {DCE-MRA}
               using inherently regularized iterative {SENSE}",
@@ -21,9 +20,7 @@
   pages    = "514--526",
   month    =  aug,
   year     =  2010
-
 }
-
 @article{Uecker:2010,
   title    = "Nonlinear inverse reconstruction for real-time {MRI} of the human
               heart using undersampled radial {FLASH}",
@@ -35,7 +32,6 @@
   month    =  jun,
   year     =  2010
 }
-
 @article{Jung:2010,
   url = {https://onlinelibrary.wiley.com/doi/10.1002/mrm.22172},
   title    = "Radial k-t {FOCUSS} for high-resolution cardiac cine {MRI}",
@@ -47,7 +43,6 @@
   month    =  jan,
   year     =  2010
 }
-
 @misc{Feng:2014,
   title   = "Golden-angle radial sparse parallel {MRI}: Combination of
              compressed sensing, parallel imaging, and golden-angle radial
@@ -61,7 +56,6 @@
   pages   = "707--717",
   year    =  2014
 }
-
 @inproceedings{tamir:2016,
   url = {http://wwwuser.gwdg.de/~muecker1/sedona16-bart.pdf},
   title={Generalized magnetic resonance image reconstruction using the Berkeley advanced reconstruction toolbox},
@@ -76,15 +70,13 @@
   booktitle={Joint Annual Meeting ISMRM-ESMRMB Meeting 2022, London, UK},
   year={2022}
 }
-
-@inproceedings{@AriyurekISMRMMoCo:2022,
+@inproceedings{AriyurekISMRMMoCo:2022,
   url = {https://submissions.mirasmart.com/ISMRM2022/Itinerary/Files/PDFFiles/0857.html},
   title={Prospective Motion Correction in Kidney MRI Using Free Induction Decay Navigators},
   author={Ariyurek, Cemre and Wallace, Tess E and Kober, Tobias and Ko\c{c}anao\u{g}ullar\i, Aziz and Kurugol, Sila and Afacan, Onur},
   booktitle={Joint Annual Meeting ISMRM-ESMRMB Meeting 2022, London, UK},
   year={2022}
 }
-
 @inproceedings{KocanaogullariISMRMMoCo:2022,
   url = {https://submissions.mirasmart.com/ISMRM2022/Itinerary/Files/PDFFiles/1958.html},
   title={Gaussian Process Modelling and Compensation of Motion in DCE-MRI},
@@ -99,13 +91,11 @@
   year = 2020,
   note = {Source code available at https://github.com/mmuckley/torchkbnufft}
 }
-
 @misc{ismrmdpytools,
   author={github},
   title={ismrmrd-python-tools},
   url={https://github.com/ismrmrd/ismrmrd-python-tools},
 }
-
 @article{coll2020bulk,
   title={Bulk motion-compensated DCE-MRI for functional imaging of kidneys in newborns},
   author={Coll-Font, Jaume and Afacan, Onur and Chow, Jeanne S and Lee, Richard S and Stemmer, Alto and Warfield, Simon K and Kurugol, Sila},
@@ -116,7 +106,6 @@
   year={2020},
   publisher={Wiley Online Library}
 }
-
 @article{coll2021modeling,
   title={Modeling dynamic radial contrast enhanced MRI with linear time invariant systems for motion correction in quantitative assessment of kidney function},
[…]
