+++ references.bib
@@ -7,7 +7,8 @@
   number   =  4,
   pages    = "1106--1115",
   month    =  apr,
-  year     =  2016
+  year     =  2016,
+  doi = {10.1109/TMI.2015.2509245 }
 }
 @article{Prieto:2010,
   title    = "{3D} undersampled golden-radial phase encoding for {DCE-MRA}
@@ -19,7 +20,8 @@
   number   =  2,
   pages    = "514--526",
   month    =  aug,
-  year     =  2010
+  year     =  2010, 
+  doi = {10.1002/mrm.22446}
 }
 @article{Uecker:2010,
   title    = "Nonlinear inverse reconstruction for real-time {MRI} of the human
@@ -30,7 +32,8 @@
   number   =  6,
   pages    = "1456--1462",
   month    =  jun,
-  year     =  2010
+  year     =  2010,
+  doi = {10.1002/mrm.22453}
 }
 @article{Jung:2010,
   url = {https://onlinelibrary.wiley.com/doi/10.1002/mrm.22172},
@@ -41,7 +44,8 @@
   number   =  1,
   pages    = "68--78",
   month    =  jan,
-  year     =  2010
+  year     =  2010,
+  doi = {10.1002/mrm.22172}
 }
 @misc{Feng:2014,
   title   = "Golden-angle radial sparse parallel {MRI}: Combination of
@@ -54,7 +58,8 @@
   volume  =  72,
   number  =  3,
   pages   = "707--717",
-  year    =  2014
+  year    =  2014,
+  doi = {10.1002/mrm.24980}
 }
 @inproceedings{tamir:2016,
   url = {http://wwwuser.gwdg.de/~muecker1/sedona16-bart.pdf},
@@ -104,7 +109,8 @@
   number={1},
   pages={207--216},
   year={2020},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1002/jmri.27021}
 }
 @article{coll2021modeling,
   title={Modeling dynamic radial contrast enhanced MRI with linear time invariant systems for motion correction in quantitative assessment of kidney function},
@@ -113,7 +119,8 @@
   volume={67},
   pages={101880},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.media.2020.101880}
 }
 @article{Ariyurek2022figshare,
 author = "Cemre Ariyurek and Sila Kurugol and Onur Afacan",
