### 2022-08-22 #4670 @cariyurek
Hi @emdupre ,

We are grateful for your time in considering and reviewing our submission. We do not have any strong suggestions for potential reviewers but looking into the list provided we can suggest curtcorum, kerkelae, spinicist or agahkarakuzu (who is not on the list but has published in JOSS).
