### 2022-10-06 #4670 @emdupre
@cariyurek , a few editorial requests :

- In your paper, can you capitalize all instances of "Python" ? It is currently lower-cased throughout the text; for example, on line 66, line 69, and line 84.
- Similarly, please capitalize mention of "Docker" on line 28.
- Please add a comma after abbreviations such as "e.g.", "i.e.", etc. This impacts line 47 and line 59, among others.
- Figure 1 renders at a low resolution. Can you include a higher-resolution of the figure in the paper materials ?
- Currently, your GitHub repository does not directly include contributing guidelines. Can you add these, for example in a CONTRIBUTING.md file ?
