+++ README.md
@@ -1,17 +1,17 @@
 # PyGRASP: A library to reconstruct Dynamic Contrast enhanced MR Images acquired with radial sampling
 
-PyGRASP is a python library for temporally regularized iterative compressed sensing reconstruction (GRASP) for dynamic contrast enhanced MRI (DCE-MRI), developed by the team at Quantitative Intelligent Imaging Lab ([QUIN](https://projects.iq.harvard.edu/quin/home)) at the radiology department of Boston Children's Hospital and Harvard Medical School. It assumes the k-space data is acquired by a Non-Cartesian, radial k-space sampling using the golden angle stack of stars trajectory.  The output is a dynamic set of volumes reconstructed at the desired temporal resolution. PyGRASP uses torchkbnufft library for a fast implementation of non-uniform Fast Fourier Transform (NUFFT) with Kaiser-Bessel gridding in PyTorch. PyGRASP uses GPU or CPU for parallelization of the NUFFT computations over the large number of dynamic volumes to be reconstructed over time.
+PyGRASP is a Python library for temporally regularized iterative compressed sensing reconstruction (GRASP) for dynamic contrast enhanced MRI (DCE-MRI), developed by the team at Quantitative Intelligent Imaging Lab ([QUIN](https://projects.iq.harvard.edu/quin/home)) at the radiology department of Boston Children's Hospital and Harvard Medical School. It assumes the k-space data is acquired by a Non-Cartesian, radial k-space sampling using the golden angle stack of stars trajectory.  The output is a dynamic set of volumes reconstructed at the desired temporal resolution. PyGRASP uses torchkbnufft library for a fast implementation of non-uniform Fast Fourier Transform (NUFFT) with Kaiser-Bessel gridding in PyTorch. PyGRASP uses GPU or CPU for parallelization of the NUFFT computations over the large number of dynamic volumes to be reconstructed over time.
 
 There are several parameters that can be modified as needed including the number of radial lines (spokes) per volume, which determines the temporal resolution and also the number of volumes generated and the regularization parameter lambda of the GRASP reconstruction that determines the amount of temporal regularization. 
 
 This library also provides the options to perform coil compression and/or coil removal before performing the reconstruction in order to speed up the reconstruction and also to improve image quality by removing coils that include severe artifacts.
 
-PyGRASP is implemented completely in Python, facilitating flexible deployment with readable code and no complation. We also provide containerized version of the library in dockers for GPU and CPU options for quick deployment.
+PyGRASP is implemented completely in Python, facilitating flexible deployment with readable code and no complation. We also provide containerized version of the library in Dockers for GPU and CPU options for quick deployment.
 
-If you use this library or its containerized docker versions please cite our [JOSS](https://joss.theoj.org/) paper.
+If you use this library or its containerized Docker versions please cite our [JOSS](https://joss.theoj.org/) paper.
 
 ## Instructions to run dce_mri
-There are three main parts - and python scripts - of the reconstruction pipeline:
+There are three main parts - and Python scripts - of the reconstruction pipeline:
 <ol>
   <li>convert_raw_to_dataset.py</li>
   Reads raw data file (currently only reads Siemens raw data), calculates coil sensitivity maps, writes k-space information and reconstruction parameters to the configuration files.
@@ -80,9 +80,9 @@ Note that spv and lam values should be entered to grasp_params.json file by the 
 ## Options for parallel processing
[…]
