### 2022-10-07 #4670 @Lestropie
It's a little out of my comfort zone, and I wouldn't be able to commence until after the [ISMRM DSGW](https://www.ismrm.org/workshops/2022/Diffusion/) next week. If you're still struggling to find better reviewers after a couple of weeks, I could have a look, but might be limited in what I can and can't be critical of. Keep me on the list of potential reviewers regardless.

Will ping a colleague who would be a good fit but doesn't appear to yet have a GitHub profile.
