+++ CONTRIBUTING.md
@@ -1,0 +1,5 @@
+# Contributing
+
+Thank you for considering contributing to pyGRASP. You can contribute by reporting an issue or by directly contributing to the source code. For the latter, fork our repository, clone the fork, make your changes, and create a pull request (PR) with a clear description of your changes -- if you are unfamiliar about forking/creating PR, please see [this guide](https://guides.github.com/activities/forking/) first. If/when your changes are merged, you will appear as one of our [Contributors](https://github.com/quin-med-harvard-edu/pyGRASP/graphs/contributors). 
+
+If you would like to report a bug, please check if similar issue have already been reported [here](https://github.com/quin-med-harvard-edu/pyGRASP/issues). If none exist please create a new issue and include as many details as possible to reproduce the issue.
