### 2022-10-13 #4670 @emdupre
Thank you @cariyurek for these edits, and @Lestropie and @spinicist for your responses !

One small clarification on your availability, @spinicist : Would you be willing and able to act as a reviewer for this submission given that the 6 week review window would run through the end of November ?

Please let us know, too, @grlee77 and @michaelberks, if you would be available to review this submission !
