### 2022-10-14 #4670 @emdupre
👋 Hi @kerkelae @curtcorum !

Would you be willing to review PyGRASP: A library to reconstruct dynamic contrast enhanced MR images acquired with radial sampling for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/quin-med-harvard-edu/pyGRASP) and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.04670/joss.04670/10.21105.joss.04670.pdf

If you're available to review this work, please let me know and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you !
