### 2022-10-26 quin-med-harvard-edu/pyGRASP#1 @kerkelae
About installation and dependencies
As a part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4865), I would like to suggest an important improvement in the `README` file. As far as I can tell, the repository does not contain an installable Python package, but a collection of scripts instead. I found a [conda environment file](https://github.com/quin-med-harvard-edu/pyGRASP/blob/main/requirements.yml) that I assume lists all the software requirements for running the scripts. Is the code meant to be run by cloning the repository, installing the requirements, and running the scripts? If so, I think that it would be important to add a description of these steps in the `README` file.
