### 2022-10-26 quin-med-harvard-edu/pyGRASP#2 @kerkelae
Where do I download the example data?
As a part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4865), I would like to ask for a clarification on how to [run the code on raw data](https://github.com/quin-med-harvard-edu/pyGRASP#demo). Where can I download the example data and what do I need to change in the mentioned files?
