### 2022-10-31 #4865 @emdupre
Hi @curtcorum 👋, I wanted to make sure you weren't facing issues in getting started since you haven't yet created your review checklist.
Please let me know if there's anything I can do to help, and thank you again for agreeing to review PyGRASP !
