### 2022-11-01 #4865 @curtcorum
I would recommend that the article be re-catagorized to: https://github.com/openjournals/joss-reviews/issues?q=is%3Aopen+label%3A%22Track%3A+2+%28BCM%29%22

instead of: https://github.com/openjournals/joss-reviews/labels/Track%3A%204%20%28SBCS%29
