### 2022-11-01 #4865 @curtcorum
The TorchKbNufft reference is incomplete: and ideally both the abstract and the github link.

Muckley, M. J., Stern, R., Murrell, T., & Knoll, F. (2020). TorchKbNufft: a high-level, hardware-agnostic non-uniform fast Fourier transform. In ISMRM Workshop on Data Sampling & Image Reconstruction.

https://github.com/mmuckley/torchkbnufft

https://mmuckley.github.io/assets/publications/2020muckleytorchkbnufft.pdf
