### 2022-11-01 #4865 @curtcorum
The text of the paper does not seem to be broken down in to reasonable sized paragraphs. Instead there is one long paragraph per section.
