### 2022-11-01 quin-med-harvard-edu/pyGRASP#2 @curtcorum
https://github.com/quin-med-harvard-edu/pyGRASP/blob/main/MRU.csv

There does not appear to be a valid link to any externally stored datafile for the demo?
