### 2022-11-01 #4865 @curtcorum
Functionality - Installation
Documentation - Installation instructions

https://github.com/quin-med-harvard-edu/pyGRASP/issues/1#issue-1423865969
