### 2022-11-01 #4865 @curtcorum
General checks - Data sharing
General checks - Reproducibility

Is the data needed for Figure 2 results (two datasets?) available?

https://github.com/quin-med-harvard-edu/pyGRASP/issues/2
