### 2022-11-01 quin-med-harvard-edu/pyGRASP#2 @curtcorum
Is the data needed for Figure 2 results (two datasets?) available?
