### 2022-11-02 quin-med-harvard-edu/pyGRASP#1 @cariyurek
Thank you for pointing this issue @kerkelae. I have realized that we have mentioned to use the environment file for the Docker but forgot to do so for running the scripts. We will add this description as suggested.
