### 2022-11-02 quin-med-harvard-edu/pyGRASP#2 @cariyurek
You may download the dataset from [Figshare](https://figshare.com/articles/dataset/DCE_MRI_MRU_dat/20465637) and this is the data for Figure 2. You need to change '.csv' file to locate the data as described in README.md. Please do not hesitate to comment in this thread if you have further questions.
