### 2022-11-06 #4865 @cariyurek
Dear @curtcorum and @kerkelae,

Thank you for your valuable comments and feedback. Currently, I am busy with the ISMRM deadline. I will go over all of your comments just after the ISMRM deadline. I am very sorry for the delay.
