+++ references.bib
@@ -133,7 +133,7 @@ doi = "10.6084/m9.figshare.20465637.v1"
 @inproceedings{inati:2014,
   title={A fast optimal method for coil sensitivity estimation and adaptive coil combination for complex images},
   author={Inati, Souheil J and Hansen, Michael S and Kellman, Peter},
-  booktitle={Proceedings of the 22nd Annual Meeting of ISMRM, Milan},
+  booktitle={In Proceedings of the 22nd Annual Meeting of ISMRM, Milan},
   pages={4407},
   year={2014}
 }
