+++ paper.md
@@ -36,20 +36,31 @@ bibliography: references.bib
 
 # Summary
 
-In dynamic contrast-enhanced (DCE) magnetic resonance imaging (MRI), a dynamic series of image stacks are acquired capturing the passage of contrast agent through the tissue compartments.  Dynamic radial imaging with a golden angle stack of stars sampling trajectory has been increasingly used to acquire DCE-MRI data due to its advantages compared to dynamic cartesian imaging, such as being robust to breathing motion, and ability to achieve high temporal resolution imaging. In order to reconstruct a dynamic series of volumes at high temporal resolution (~3 seconds/volume), the data needs to be reconstructed from undersampled measurements. Different regularizers have been proposed to reduce the incoherent undersampling artifacts over the dynamic image series in radial acquisition, which are in the form of streaking artifacts [@Jung:2010; @Uecker:2010; @Poddar:2016; @Prieto:2010]. Golden-angle RAdial Sparse Parallel (GRASP) [@Feng:2014] reconstruction algorithm uses temporal regularization to reduce artifacts and can reconstruct a dynamic series of volumes at the desired temporal resolution.  In this software, we have developed an efficient open-source, purely Python-based GRASP reconstruction library called pyGRASP that allows researchers to access the source code for development, facilitating flexible deployment with readable code and no compilation. This library includes preprocessing by extracting information from raw data acquired in Siemens MRI scanners and coil sensitivity map estimation, coil compression and coil removal options for improving speed and image quality and image reconstruction using a compressed sensing approach based on GRASP technique, followed by post-processing. The proposed library has options to accelerate computations by parallel processing using either GPU or CPU resources. We also provide Docker files for quick and easy deployment over any operating system or over a cloud.
+In dynamic contrast-enhanced (DCE) magnetic resonance imaging (MRI), a dynamic series of image stacks are acquired capturing the passage of contrast agent through the tissue compartments.  Dynamic radial imaging with a golden angle stack of stars sampling trajectory has been increasingly used to acquire DCE-MRI data due to its advantages compared to dynamic cartesian imaging, such as being robust to breathing motion, and ability to achieve high temporal resolution imaging. In order to reconstruct a dynamic series of volumes at high temporal resolution (~3 seconds/volume), the data needs to be reconstructed from undersampled measurements. Different regularizers have been proposed to reduce the incoherent undersampling artifacts over the dynamic image series in radial acquisition, which are in the form of streaking artifacts [@Jung:2010; @Uecker:2010; @Poddar:2016; @Prieto:2010]. Golden-angle RAdial Sparse Parallel (GRASP) [@Feng:2014] reconstruction algorithm uses temporal regularization to reduce artifacts and can reconstruct a dynamic series of volumes at the desired temporal resolution.  
+
+In this software, we have developed an efficient open-source, purely Python-based GRASP reconstruction library called pyGRASP that allows researchers to access the source code for development, facilitating flexible deployment with readable code and no compilation. This library includes preprocessing by extracting information from raw data acquired in Siemens MRI scanners and coil sensitivity map estimation, coil compression and coil removal options for improving speed and image quality and image reconstruction using a compressed sensing approach based on GRASP technique, followed by post-processing. The proposed library has options to accelerate computations by parallel processing using either GPU or CPU resources. We also provide Docker files for quick and easy deployment over any operating system or over a cloud.
 
 # Statement of need
 
[…]
