+++ references.bib
@@ -97,7 +97,7 @@
   note = {Source code available at https://github.com/mmuckley/torchkbnufft}
 }
 @misc{ismrmdpytools,
-  author={github},
+  author={ismrmrd},
   title={ismrmrd-python-tools},
   url={https://github.com/ismrmrd/ismrmrd-python-tools},
 }
@@ -136,4 +136,19 @@ doi = "10.6084/m9.figshare.20465637.v1"
   booktitle={In Proceedings of the 22nd Annual Meeting of ISMRM, Milan},
   pages={4407},
   year={2014}
+}
+@misc{github:torchkbnufft,
+  author={mmuckley},
+  title={torchkbnufft},
+  url={https://github.com/mmuckley/torchkbnufft},
+}
+@misc{github:sigPy,
+  author={mikgroup},
+  title={SigPy},
+  url={https://github.com/mikgroup/sigpy},
+}
+@misc{github:extreme_mri,
+  author={mikgroup},
+  title={Extreme MRI},
+  url={https://github.com/mikgroup/extreme_mri},
 }
