### 2022-12-01 #4865 @emdupre
Hi @cariyurek ! Now that we're on the other side of the ISMRM deadline, I just wanted to check in and see how this was going for you.

Please let me know if you have any questions in addressing @curtcorum and @kerkelae's comments !
