### 2022-12-01 aada440
Adding steps to run pyGRASP demo
