+++ requirements.yml
@@ -1,106 +1,0 @@
-name: dce_env
-channels:
-  - owlas
-  - conda-forge
-  - anaconda
-  - defaults
-dependencies:
-  - _libgcc_mutex=0.1=main
-  - blas=1.0=mkl
-  - ca-certificates=2020.10.14=0
-  - cached-property=1.5.2=py_0
-  - certifi=2020.6.20=py36_0
-  - cloudpickle=1.6.0=py_0
-  - cytoolz=0.11.0=py36h7b6447c_0
-  - dask-core=2.30.0=py_0
-  - dbus=1.13.18=hb2f20db_0
-  - decorator=4.4.2=py_0
-  - expat=2.2.10=he6710b0_2
-  - fontconfig=2.13.0=h9420a91_0
-  - freetype=2.10.4=h5ab3b9f_0
-  - glib=2.56.2=hd408876_0
-  - gst-plugins-base=1.14.0=hbbd80ab_1
-  - gstreamer=1.14.0=hb453b48_1
-  - icu=58.2=he6710b0_3
-  - imageio=2.9.0=py_0
-  - intel-openmp=2020.2=254
-  - joblib=0.17.0=py_0
-  - jpeg=9b=habf39ab_1
-  - lcms2=2.11=h396b838_0
-  - libedit=3.1.20191231=h14c3975_1
-  - libffi=3.2.1=hf484d3e_1007
-  - libgcc=7.2.0=h69d50b8_2
-  - libgcc-ng=9.1.0=hdf63c60_0
-  - libgfortran-ng=7.3.0=hdf63c60_0
-  - libpng=1.6.37=hbc83047_0
-  - libstdcxx-ng=9.1.0=hdf63c60_0
-  - libtiff=4.1.0=h2733197_1
-  - libuuid=1.0.3=h1bed415_2
-  - libxcb=1.14=h7b6447c_0
-  - libxml2=2.9.10=hb55368b_3
-  - lz4-c=1.9.2=heb0550a_3
-  - matplotlib=3.3.1=0
-  - matplotlib-base=3.3.1=py36h817c723_0
-  - mkl=2020.2=256
-  - mkl-service=2.3.0=py36he8ac12f_0
-  - mkl_fft=1.2.0=py36h23d657b_0
-  - mkl_random=1.1.1=py36h0573a6f_0
-  - ncurses=6.2=he6710b0_1
-  - networkx=2.5=py_0
-  - numpy-base=1.19.2=py36hfa32c7d_0
-  - olefile=0.46=py36_0
-  - openssl=1.0.2u=h7b6447c_0
-  - pandas=1.1.3=py36he6710b0_0
-  - pcre=8.44=he6710b0_0
-  - pip=20.3.3=py36h06a4308_0
-  - pyparsing=2.4.7=py_0
-  - pyqt=5.9.2=py36h22d08a2_1
-  - python=3.6.5=hc3d631a_2
-  - python-dateutil=2.8.1=pyhd3eb1b0_0
-  - pytz=2021.1=pyhd3eb1b0_0
-  - pywavelets=1.1.1=py36h7b6447c_2
-  - pyyaml=5.3.1=py36h7b6447c_1
-  - qt=5.9.6=h8703b6f_2
-  - readline=7.0=h7b6447c_5
-  - scikit-image=0.17.2=py36hdf5156a_0
-  - setuptools=52.0.0=py36h06a4308_0
-  - sip=4.19.24=py36he6710b0_0
-  - six=1.15.0=py36h06a4308_0
-  - sqlite=3.33.0=h62c20be_0
-  - tifffile=2020.10.1=py36hdd07704_2
-  - tk=8.6.10=hbc83047_0
-  - toolz=0.11.1=py_0
[…]
