### 2022-12-01 f1e5ecf
Adding steps to run pyGRASP demo
