+++ README.md
@@ -89,14 +89,22 @@ In order to start running dce_mri on a Docker container, your current directory 
 You may run the code by cloning the repository, installing the environment, and running the scripts. An example script "pyGRASP_demo.sh" has been provided to run a demo case along with the raw data. Environment, code and input-output paths have to be edited in pyGRASP_demo.sh. Also, raw data address has to be updated in MRU.csv. Then, one can run the demo via "bash pyGRASP_demo.sh" command on terminal. Demo data DCE_MRI_MRU.dat can be downloaded from figshare[@Ariyurek2022figshare]. 
 
 Steps to run pyGRASP demo:
-    1) Download the data figshare[@Ariyurek2022figshare]
-    2) Download/clone the repository pyGRASP
-    3) Create the environment  from environment.yml (e.g., "conda env create -f environment.yml" in command line)
-    4) Dedicate a folder for inputOutput_path
-    5) Copy MRU.csv to inputOutput_path folder and edit it by replacing the directory path of the demo data 
-    6) Edit the environment, code, input/output paths in pyGRASP_demo.sh 
-    7) Run the pyGRASP_demo.sh (e.g., "bash pyGRASP_demo.sh" in command line)
-    8) You may find the reconstructed 4D DCE-MRI result as a NIfTI file in "inputOutput_path/../sub-1/date_time-rec4D.nii.gz" along with the JSON file which records the parameters used in the reconstruction.  
+    
+1) Download the data figshare[@Ariyurek2022figshare]
+    
+2) Download/clone the repository pyGRASP
+    
+3) Create the environment  from environment.yml (e.g., "conda env create -f environment.yml" in command line)
+    
+4) Dedicate a folder for inputOutput_path
+    
+5) Copy MRU.csv to inputOutput_path folder and edit it by replacing the directory path of the demo data 
+    
+6) Edit the environment, code, input/output paths in pyGRASP_demo.sh 
+    
+7) Run the pyGRASP_demo.sh (e.g., "bash pyGRASP_demo.sh" in command line)
+    
+8) You may find the reconstructed 4D DCE-MRI result as a NIfTI file in "inputOutput_path/../sub-1/*date_time*-rec4D.nii.gz" along with the JSON file which records the parameters used in the reconstruction.  
 
 Alternative to command prompt you can the code in a python editor/compiler. 
 
