### 2022-12-01 #4865 @cariyurek
> Software Paper - References
>
> The TorchKbNufft reference is incomplete, and ideally should include abstract and the github link.
>
> Muckley, M. J., Stern, R., Murrell, T., & Knoll, F. (2020). TorchKbNufft: a high-level, hardware-agnostic non-uniform fast Fourier transform. In ISMRM Workshop on Data Sampling & Image Reconstruction.
>
> https://github.com/mmuckley/torchkbnufft
>
> https://mmuckley.github.io/assets/publications/2020muckleytorchkbnufft.pdf

Thank you for pointing this out. I have added the suggested references to the paper.
