### 2022-12-01 #4865 @cariyurek
> Software Paper - Quality of writing
>
> The text of the paper does not seem to be broken down in to reasonable sized paragraphs. Instead there is one long paragraph per section.

We have divided sections into paragraphs to ease the readability as suggested.
