### 2022-12-01 #4865 @cariyurek
> Software Paper - References
>
> SigPy reference or discussion? Sigpy is a general tool that in principle can perform GRASP in a Python based environment. There is not an explicit tutorial or dateset available as far as this reviewer can tell, though.
>
> https://github.com/mikgroup/sigpy https://github.com/mikgroup/extreme_mri

We have added some text at the end of statement of need section discussing and citing SigPy tool.
