### 2022-12-01 #4865 @cariyurek
> General checks - Data sharing General checks - Reproducibility
>
> Is the data needed for Figure 2 results (two datasets?) available?
>
> [quin-med-harvard-edu/pyGRASP#2](https://github.com/quin-med-harvard-edu/pyGRASP/issues/2)

Yes, you can generate the results in Figure 2 by running the demo for different regularization parameters. The data is avaliable in [figshare](https://figshare.com/articles/dataset/DCE_MRI_MRU_dat/20465637). We have added steps to run the demo to the demo section in readme.md.
