### 2022-12-01 quin-med-harvard-edu/pyGRASP#1 @cariyurek
I have added steps to run the demo using the data provided in [figshare](https://figshare.com/articles/dataset/DCE_MRI_MRU_dat/20465637) to the demos section in readme.md. I have tried to run the demo from the scratch by cloning the repository and installing the environment. I have realized that requirements.yml file had some missing packages and [environment.yml](https://github.com/quin-med-harvard-edu/pyGRASP/blob/main/environment.yml) was the complete one, so I removed the requirements.yml file.

Please let me know if you have any trouble running the demo.
