### 2022-12-06 #4865 @emdupre
Hi @cariyurek , thank you for making these changes based on reviewer feedback  !

Please let us know, @curtcorum and @kerkelae, if the changes significantly address your concerns.

I also did want to note on your suggestion for track assignment, @curtcorum : this is well-received, thank you ! The primary purpose of track assignments (at this point) is to help improving the editorial assignment of papers, so I think changing this won't impact the paper's status in any sense. But I appreciate your attention to this detail !
