### 2022-12-07 quin-med-harvard-edu/pyGRASP#2 @cariyurek
> Hi @cariyurek , thank you for this clarification !
>
> Looking at the [demo](https://github.com/quin-med-harvard-edu/pyGRASP#demo) in the README, the data availability is not immediately obvious as the figshare link renders as:
>
> ```
> Download the data figshare[@Ariyurek2022figshare]
> ```
>
> Could you please update this to include a resolvable hyperlink ?
