### 2022-12-07 quin-med-harvard-edu/pyGRASP#2 @cariyurek
Hi @emdupre ,

Thank you for pointing this out. I have updated the "figshare" as a resolvable link.
