### 2022-12-14 #4865 @kerkelae
The first author (@cariyurek) is the only one who has made commits in the repository. Can you please describe how the other authors have contributed and if they deserve authorship according to the [JOSS policy](https://joss.readthedocs.io/en/latest/submitting.html#authorship)?
