### 2022-12-14 quin-med-harvard-edu/pyGRASP#1 @kerkelae
Hi @cariyurek, I am unable to run the demo:

```
(dce_env) leevi@leevi-laptop:~/Desktop/joss-review/pyGRASP$ bash pyGRASP_demo.sh
  0%|                                                                                                                                                                                        | 0/1 [00:00<?, ?it/s]pymapVBVD version 0.4.1
Software version: VD
Scan 1/2, read all mdhs: 57.0MB [00:00, 350MB/s]
Scan 1/2, read all mdhs:  73%|██████████████████████████████████████████████████████████████████████████████████████████████████████████▍                                      | 41.8M/57.0M [00:00<00:00, 438MB/s/home/leevi/anaconda3/envs/dce_env/lib/python3.7/site-packages/mapvbvd/twix_map_obj.py:755: RuntimeWarning: invalid value encountered in add█████████████████████████████████▉ | 11.2G/11.3G [00:11<00:00, 2.12GB/s]
  raw = (raw[:, 0] + 1j * raw[:, 1]).reshape(readShape, order='F')                                                                                                         | 9342/111384 [00:07<01:15, 1354.21it/s]
                                                                                                                                                                                                                  pyGRASP_demo.sh: line 5: 53012 Killed                  "$env_path"/dce_env/bin/python3.7"" "$code_path"/convert_raw_to_dataset.py"" -i "$inputOutput_path"/MRU.csv"" -pg "$code_path"/parameters/grasp_params.json"" -pp "$code_path"/parameters/p_proc_params.json"" -o $inputOutput_path -fr True -fc True -nspkc 500
Traceback (most recent call last):
  File "/home/leevi/Desktop/joss-review/pyGRASP/demo_reconstruction.py", line 70, in <module>
    crl_msu_dat = CRLMRUData(root_csv_file, type_parse='slice')
  File "/home/leevi/Desktop/joss-review/pyGRASP/data_loader/crl_dataset.py", line 21, in __init__
    with open(self.csv_root, 'r') as file:
FileNotFoundError: [Errno 2] No such file or directory: '/home/leevi/Desktop/joss-review/inputOutput/loader.csv'
Traceback (most recent call last):
  File "/home/leevi/Desktop/joss-review/pyGRASP/slice_to_vol.py", line 76, in <module>
    for rec_file in os.listdir(folder_raw_rec):
FileNotFoundError: [Errno 2] No such file or directory: '/home/leevi/Desktop/joss-review/inputOutput/../sub-1/*/raw-rec/'
```
Do you know what may be causing this?
