### 2022-12-15 quin-med-harvard-edu/pyGRASP#1 @cariyurek
Hi, @kerkelae ,

I am sorry to hear that. I tried to replicate the error but I could not.

You are getting an error when data is read. I suspect that pymapvbvd does not have the correct version installed. Could you run "conda list" and send me the output?

Also, could you be sure that your environment is pointed correctly?
