### 2022-12-20 quin-med-harvard-edu/pyGRASP#1 @kerkelae
Below is the contents of `pyGRASP_demo.sh`:

```
env_path="/home/leevi/anaconda3/envs"
code_path="/home/leevi/Desktop/joss-review/pyGRASP"
inputOutput_path="/home/leevi/Desktop/joss-review/inputOutput"

"$env_path"/dce_env/bin/python3.7"" "$code_path"/convert_raw_to_dataset.py"" -i "$inputOutput_path"/MRU.csv"" -pg "$code_path"/parameters/grasp_params.json"" -pp "$code_path"/parameters/p_proc_params.json"" -o $inputOutput_path -fr True -fc True -nspkc 500

export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:"$env_path"/dce_env/lib""

"$env_path"/dce_env/bin/python3.7"" "$code_path"/demo_reconstruction.py"" -p "$code_path"/parameters/grasp_params.json"" -d "$inputOutput_path"/loader.csv"" -o . -sid sub-1

"$env_path"/dce_env/bin/python3.7"" "$code_path"/slice_to_vol.py"" -p "$code_path"/parameters/p_proc_params.json"" -d "$inputOutput_path"/../sub-1/*/raw-rec/"" -o "$inputOutput_path"/../sub-1/""
```
And below is what `conda list` gives:
```
# packages in environment at /home/leevi/anaconda3/envs/dce_env:
#
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                 conda_forge    conda-forge
_openmp_mutex             4.5                       1_gnu    conda-forge
attrdict                  2.0.1                      py_0    conda-forge
attrs                     20.3.0                   pypi_0    pypi
backcall                  0.2.0                      py_0    anaconda
blas                      1.0                         mkl    anaconda
bzip2                     1.0.8                h7f98852_4    conda-forge
ca-certificates           2021.10.26           h06a4308_2    anaconda
cached-property           1.5.2                      py_0    anaconda
certifi                   2021.10.8        py37h06a4308_0    anaconda
charset-normalizer        2.0.8                    pypi_0    pypi
chart-studio              1.1.0                    pypi_0    pypi
cloudpickle               1.6.0                      py_0    anaconda
cycler                    0.10.0                     py_2    conda-forge
cytoolz                   0.11.0           py37h5e8e339_3    conda-forge
dask-core                 2.30.0                     py_0    anaconda
dataclasses               0.6                      pypi_0    pypi
dbus                      1.13.18              hb2f20db_0    anaconda
decorator                 4.4.2                      py_0    anaconda
expat                     2.2.10               he6710b0_2    anaconda
fontconfig                2.13.0               h9420a91_0    anaconda
freetype                  2.10.4               h5ab3b9f_0    anaconda
glib                      2.56.2               hd408876_0    anaconda
gst-plugins-base          1.14.0               hbbd80ab_1    anaconda
gstreamer                 1.14.0               hb453b48_1    anaconda
h5py                      3.1.0                    pypi_0    pypi
hdf5storage               0.1.18                   pypi_0    pypi
icu                       58.2                 he6710b0_3    anaconda
idna                      3.3                      pypi_0    pypi
imagecodecs-lite          2019.12.3        py37h902c9e0_3    conda-forge
imageio                   2.9.0                      py_0    anaconda
importlib-metadata        3.10.0                   pypi_0    pypi
iniconfig                 1.1.1                    pypi_0    pypi
intel-openmp              2020.2                      254    anaconda
ipykernel                 5.3.4            py37h5ca1d4c_0    anaconda
ipython                   7.18.1           py37h5ca1d4c_0    anaconda
ipython_genutils          0.2.0                    py37_0    anaconda
jedi                      0.17.2                   py37_0    anaconda
joblib                    0.17.0                     py_0    anaconda
jpeg                      9b                   habf39ab_1    anaconda
jupyter_client            6.1.7                      py_0    anaconda
jupyter_core              4.6.3                    py37_0    anaconda
kiwisolver                1.3.1            py37hc928c03_0    conda-forge
lcms2                     2.11                 h396b838_0    anaconda
libedit                   3.1.20191231         h14c3975_1    anaconda
libffi                    3.2.1             hf484d3e_1007    anaconda
libgcc                    7.2.0                h69d50b8_2    anaconda
libgcc-ng                 9.3.0               h2828fa1_18    conda-forge
libgfortran-ng            7.3.0                hdf63c60_0    anaconda
libgomp                   9.3.0               h2828fa1_18    conda-forge
libpng                    1.6.37               hbc83047_0    anaconda
libsodium                 1.0.18               h7b6447c_0    anaconda
libstdcxx-ng              9.1.0                hdf63c60_0    anaconda
libtiff                   4.1.0                h2733197_1    anaconda
libuuid                   1.0.3                h1bed415_2    anaconda
libxcb                    1.14                 h7b6447c_0    anaconda
libxml2                   2.9.10               hb55368b_3    anaconda
lz4-c                     1.9.2                heb0550a_3    anaconda
matplotlib                3.3.2                    pypi_0    pypi
mkl                       2020.2                      256    anaconda
mkl-service               2.3.0            py37h8f50634_2    conda-forge
mkl_fft                   1.2.0            py37h161383b_1    conda-forge
mkl_random                1.2.0            py37h9fdb41a_1    conda-forge
ncurses                   6.2                  he6710b0_1    anaconda
networkx                  2.5                        py_0    anaconda
nibabel                   3.2.1                    pypi_0    pypi
numpy                     1.19.2           py37h54aff64_0
numpy-base                1.19.2           py37hfa32c7d_0
olefile                   0.46               pyh9f0ad1d_1    conda-forge
openssl                   1.1.1l               h7f8727e_0    anaconda
packaging                 20.9                     pypi_0    pypi
pandas                    1.0.1                    pypi_0    pypi
parso                     0.7.0                      py_0    anaconda
patsy                     0.5.2                    pypi_0    pypi
pcre                      8.44                 he6710b0_0    anaconda
pexpect                   4.8.0                    py37_1    anaconda
pickleshare               0.7.5                 py37_1001    anaconda
pillow                    8.2.0            py37he98fc37_0    anaconda
pip                       21.0.1             pyhd8ed1ab_0    conda-forge
plotly                    5.4.0                    pypi_0    pypi
pluggy                    0.13.1                   pypi_0    pypi
prompt-toolkit            3.0.8                      py_0    anaconda
psutil                    5.8.0                    pypi_0    pypi
ptyprocess                0.6.0                    py37_0    anaconda
py                        1.10.0                   pypi_0    pypi
pygments                  2.7.1                      py_0    anaconda
pymapvbvd                 0.4.1                    pypi_0    pypi
pynrrd                    0.4.2                    pypi_0    pypi
pynufft                   2021.0.1                 pypi_0    pypi
pyparsing                 2.4.7                      py_0    anaconda
pyqt                      5.9.2            py37hcca6a23_4    conda-forge
pytest                    6.2.3                    pypi_0    pypi
python                    3.7.3                h5b0a415_0    conda-forge
python-dateutil           2.8.1              pyhd3eb1b0_0    anaconda
python_abi                3.7                     1_cp37m    conda-forge
pytz                      2021.1             pyhd3eb1b0_0    anaconda
pywavelets                1.1.1            py37h902c9e0_3    conda-forge
pyyaml                    5.4.1            py37h5e8e339_0    conda-forge
pyzmq                     19.0.2           py37he6710b0_1    anaconda
qt                        5.9.7                h5867ecd_1    anaconda
readline                  7.0                  h7b6447c_5    anaconda
requests                  2.26.0                   pypi_0    pypi
retrying                  1.3.3                    pypi_0    pypi
scikit-image              0.18.2                   pypi_0    pypi
scikit-learn              0.24.1                   pypi_0    pypi
scipy                     1.5.4                    pypi_0    pypi
setuptools                49.6.0           py37h89c1867_3    conda-forge
simpleitk                 2.0.2                    pypi_0    pypi
sip                       4.19.8          py37hf484d3e_1000    conda-forge
six                       1.15.0             pyh9f0ad1d_0    conda-forge
sqlite                    3.33.0               h62c20be_0    anaconda
statsmodels               0.13.1                   pypi_0    pypi
tenacity                  8.0.1                    pypi_0    pypi
threadpoolctl             2.1.0              pyh5ca1d4c_0    anaconda
tifffile                  2020.6.3                   py_0    conda-forge
tk                        8.6.10               hbc83047_0    anaconda
toml                      0.10.2                   pypi_0    pypi
toolz                     0.11.1                     py_0    anaconda
torch                     1.8.1                    pypi_0    pypi
torchkbnufft              1.1.0                    pypi_0    pypi
tornado                   6.1              py37h5e8e339_1    conda-forge
tqdm                      4.56.0             pyhd8ed1ab_0    conda-forge
traitlets                 5.0.5                      py_0    anaconda
transforms3d              0.3.1                h2e23829_0    owlas
typing-extensions         3.7.4.3                  pypi_0    pypi
urllib3                   1.26.7                   pypi_0    pypi
wcwidth                   0.2.5                      py_0    anaconda
webcolors                 1.11.1                   pypi_0    pypi
wheel                     0.36.2             pyhd3eb1b0_0    anaconda
xlrd                      2.0.1              pyhd8ed1ab_3    conda-forge
xz                        5.2.5                h7b6447c_0    anaconda
yaml                      0.2.5                h7b6447c_0    anaconda
zeromq                    4.3.3                he6710b0_3    anaconda
zipp                      3.4.1                    pypi_0    pypi
zlib                      1.2.11               h7b6447c_3    anaconda
zstd                      1.4.4                h0b5b093_3    anaconda
```
