### 2022-12-27 quin-med-harvard-edu/pyGRASP#1 @cariyurek
Hi @kerkelae ,

Your environment list seems correct.

Could you delete everything in the folder "inputOutput_path" except MRU.csv. and just run the data reader part:

env_path="/home/leevi/anaconda3/envs"
code_path="/home/leevi/Desktop/joss-review/pyGRASP"
inputOutput_path="/home/leevi/Desktop/joss-review/inputOutput"

"$env_path"/dce_env/bin/python3.7"" "$code_path"/convert_raw_to_dataset.py"" -i "$inputOutput_path"/MRU.csv"" -pg "$code_path"/parameters/grasp_params.json"" -pp "$code_path"/parameters/p_proc_params.json"" -o $inputOutput_path -fr True -fc True -nspkc 500

You can comment or delete the rest of pyGRASP_demo.sh for now just to debug this part. And could you send me the error?

Sorry for the inconvenience again.
