### 2022-12-28 #4865 @cariyurek
> The first author (@cariyurek) is the only one who has made commits in the repository. Can you please describe how the other authors have contributed and if they deserve authorship according to the [JOSS policy](https://joss.readthedocs.io/en/latest/submitting.html#authorship)?

This repository was copied from our main private repository. Some of the features in that repository were not ready to share (i.e., work in progress). So, I created a new repository for this submission. The authors: Aziz Kocanaogullari, Can Taylan Sari and Serge Vasylechko contributed to code and building Docker; other authors: Onur Afacan and Sila Kurugol contributed to the idea, design and writing the paper as PIs. Hence, commits made by other authors are available in our private main repository but not in this repository.
