### 2023-01-03 #4865 @emdupre
Hi @cariyurek,

Happy New Year, and thank you for your engagement with this review process to date ! I just wanted to take a moment to summarize my understanding of where this review currently sits and outline the current actionables to move forward.

Both reviewers have noted difficulty in installing PyGRASP, as outlined in https://github.com/quin-med-harvard-edu/pyGRASP/issues/1 . In general, JOSS recommends that all installation requirements are ideally handled with an automated package management solution [according to common community standards for the programming language being used](https://joss.readthedocs.io/en/latest/submitting.html#). For Python, this is primarily through `pip`; indeed, I think moving to a pip managed installation could resolve most of the installation issues encountered to date. Current best practices for this are outlined in the [Python packaging guide](https://packaging.python.org/en/latest), particularly in [this section](https://packaging.python.org/en/latest/tutorials/packaging-projects).

At present, then, it seems that creating an installable environment for PyGRASP will significantly speed up the rest of the review process by allowing @curtcorum and @kerkelae to check the core functional claims of the software.  I would recommend restructuring the package to meet this need.

Although I have not seen this discussed, I just want to note : the Docker image files--while very useful !--cannot replace a direct installation. Asking users to install software outside of containers often exposes complex bugs worth fixing.

Please let me know if you have any questions, and thank you again for your submission !
