+++ README.md
@@ -88,13 +88,13 @@ In order to start running dce_mri on a Docker container, your current directory 
 
 You may run the code by cloning the repository, installing the environment, and running the scripts. An example script "pyGRASP_demo.sh" has been provided to run a demo case along with the raw data. Environment, code and input-output paths have to be edited in pyGRASP_demo.sh. Also, raw data address has to be updated in MRU.csv. Then, one can run the demo via "bash pyGRASP_demo.sh" command on terminal. Demo data DCE_MRI_MRU.dat can be downloaded from [figshare](https://figshare.com/articles/dataset/DCE_MRI_MRU_dat/20465637). 
 
-Steps to run pyGRASP demo:
+Steps to run pyGRASP demo using conda environment:
     
 1) Download the data [figshare](https://figshare.com/articles/dataset/DCE_MRI_MRU_dat/20465637)
     
 2) Download/clone the repository pyGRASP
     
-3) Create the environment  from environment.yml (e.g., "conda env create -f environment.yml" in command line)
+3) Create the environment from environment.yml (e.g., "conda env create -f environment.yml" in command line)
     
 4) Dedicate a folder for inputOutput_path
     
@@ -105,6 +105,30 @@ Steps to run pyGRASP demo:
 7) Run the pyGRASP_demo.sh (e.g., "bash pyGRASP_demo.sh" in command line)
     
 8) You may find the reconstructed 4D DCE-MRI result as a NIfTI file in "inputOutput_path/../sub-1/*date_time*-rec4D.nii.gz" along with the JSON file which records the parameters used in the reconstruction.  
+
+Alternatively, steps to run pyGRASP demo using pip environment:
+    
+1) Download the data [figshare](https://figshare.com/articles/dataset/DCE_MRI_MRU_dat/20465637)
+    
+2) Download/clone the repository pyGRASP
+    
+3) Create the environment from pip_requirements.txt and activate: 
+    # create virtual environment
+    python3 -m venv $PWD/venv2
+
+    # activate
+    source venv2/bin/activate
+    
+4) Dedicate a folder for inputOutput_path
+    
+5) Copy MRU.csv to inputOutput_path folder and edit it by replacing the directory path of the demo data 
+    
+6) Edit inputOutput_path in pip_demo.sh
+    
+7) Go to pyGRASP-main folder and run pip_demo.sh (e.g., "bash pip_demo.sh" in command line)
+    
+8) You may find the reconstructed 4D DCE-MRI result as a NIfTI file in "inputOutput_path/../sub-1/*date_time*-rec4D.nii.gz" along with the JSON file which records the parameters used in the reconstruction.  
+
 
 Alternative to command prompt you can the code in a python editor/compiler. 
 
