### 2023-01-05 4c7a31d
Adding steps to run pyGRASP demo with pip install environment as suggested by the editor
