+++ pip_requirements.txt
@@ -1,0 +1,10 @@
+numpy==1.22
+scipy==1.8
+hdf5storage==0.1.15
+joblib==0.17.0
+nibabel==3.2.0
+pynufft==2020.2.7
+torch==1.13.0
+torchkbnufft==1.4.0
+tqdm==4.51.0
+pymapvbvd
