### 2023-01-06 ecfb233
Adding steps to run pyGRASP demo with pip install environment as suggested by the editor
