### 2023-01-06 #4865 @cariyurek
Hi @emdupre,

Happy new year! Thank you for your effort and time with this review process.

As suggested, we have made the necessary changes for users to install packages through pip alternative to conda environment and added necessary steps to use it.

For the problem outlined by @kerkelae  in [quin-med-harvard-edu/pyGRASP#1](https://github.com/quin-med-harvard-edu/pyGRASP/issues/1), we suspect it is a memory issue while reading the data. We will provide an additional sample data (smaller in size) and ask the reviewers to use a computer with a larger memory.
