### 2023-01-06 75edfec
Adding steps to run pyGRASP demo with pip install environment as suggested by the editor
