### 2023-01-06 quin-med-harvard-edu/pyGRASP#1 @cariyurek
Hi @kerkelae ,

I was able to replicate the problem you are having on a laptop with 16GB RAM. I think you are having memory issues. We have been running pyGRASP in servers with at least 128GB RAM. It would be great if you could try to run pyGRASP on a server with a large memory. In addition, we will provide another test data (smaller in size) after de-identifying and trying to run pyGRASP on the laptop again. I will post here when we upload the new data to figshare. We are sorry for this issue.
