### 2023-01-09 quin-med-harvard-edu/pyGRASP#1 @kerkelae
Good that you were able to figure out the cause of the problem! Please let me know when the new dataset is available. I would like to be able to conduct this review using my personal computer with 32 GB of RAM. It'd also be great if you implemented a feature to let the user know when the software fails due to insufficient memory.
