### 2023-02-07 ee18814
Adding a second data to figshare that is smaller in size
