### 2023-02-08 8484ea2
Printing size of the available memory and the data, in addition to a warning about insufficient memory
