### 2023-02-08 quin-med-harvard-edu/pyGRASP#1 @cariyurek
Hi @kerkelae ,

We have provided another data that is 3 times smaller in size in [figshare](https://figshare.com/articles/dataset/DCE-MRI_raw_data/22043195) and updated demo steps in our GitHub accordingly. The new data has less number of channel measurements and poorer in quality, so you will not end up with good quality DCE-MR images as in Figure 2 in the paper. Unfortunately, I was not able to run it on a laptop with 16GB RAM. Maybe it would be possible for you to run it on your personal computer with 32 GB of RAM. Now, it prints out the data size, available memory and warning about monitoring insufficient memory.

If this does not work we may have to collect some data (not DCE-MRI but with the same MRI sequence and less number of channels) very small in size for you to run the demo.

I hope this works.
