### 2023-02-22 #4865 @cariyurek
Hi @arfon,

Yes, there was a memory issue. Therefore, we have provided another sample data that is smaller in size (poorer in quality too), which is available in [figshare_data_2](https://figshare.com/articles/dataset/DCE-MRI_raw_data/22043195). Reviewers could try to use the new sample data to test the software. Nevertheless, at least 16 GB of free memory is still required for the new data. DCE-MRI data consists of a long scan and multiple coils, hence the data is large in size. And this new data was the smallest we could find. If this one does not work too,  we need to acquire some new smaller data that is not an actual DCE-MRI protocol.

Also, we have made the necessary changes for users to install packages through pip alternative to conda environment.

We believe these edits helped the software become more 'installable'. It would be great if they could try to test it again. We are sorry for the issues with testing the software and grateful for editors' and reviewers' time in considering and reviewing this paper & software.
