### 2023-02-23 #4865 @emdupre
Thank you for confirming, @cariyurek !

I've spoken to @kerkelae out-of-band and he's confirmed that he's planning to re-assess the functionality of the software (per the [JOSS reviewing guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#functionality)) with the new demo dataset next week.

@curtcorum , please let us know if you similarly expect to be able to continue your review with this new sample dataset, or if you are encountering any other difficulties / have any specific questions at this point !
