### 2023-02-25 #4865 @curtcorum
@emdupre

I took a brief look at the updated [instuctions](https://github.com/quin-med-harvard-edu/pyGRASP/blob/4a8f0f5f71979e9b98685aef1045d34e3cc7627d/README.md), which seem much improved and will give it a try with the new dataset.

> @curtcorum , please let us know if you similarly expect to be able to continue your review with this new sample dataset, or if you are encountering any other difficulties / have any specific questions at this point !
