### 2023-03-08 quin-med-harvard-edu/pyGRASP#4 @kerkelae
Issue with reconstructed image
I was able to run the demo on my computer using the smaller dataset (although it took 8 h), but the reconstructed image (`03_07_2023_19_48_28-rec4D.nii.gz`) does not look right. Below is what I see with fsleyes. Any idea what went wrong and how to fix the issue?

![reconstruction](https://user-images.githubusercontent.com/15076450/223705018-37a57004-c161-48e3-84dd-55d7d683a5a7.png)
