### 2023-03-08 #4865 @kerkelae
Hi @emdupre, I was able to run the demo using the smaller dataset, but the reconstructed image does not look right ([GitHub issue](https://github.com/quin-med-harvard-edu/pyGRASP/issues/4)). I have time to continue with the review now.
