### 2023-03-09 quin-med-harvard-edu/pyGRASP#4 @cariyurek
Hi @kerkelae ,

I am happy to hear that you were able to run the demo. In this image, you are looking at the very first slice (back of the subject) at the very beginning of the sequence. Could you move to a slice where you could see kidneys and a mid-point in the time dimension? But it is true that this data was acquired with poor quality with some artifacts unlike the actual data (the first dataset) we shared, which produces the images in the paper.
