### 2023-03-10 9074925
If there is no slice oversampling slice_os in pproc_params is 1.0
