### 2023-03-13 quin-med-harvard-edu/pyGRASP#4 @kerkelae
Thanks @cariyurek! The coronal slices of most of the volumes look as expected. Is it expected for the sagittal and axial slices to look like this? What is causing it? Is the ordering of the slices correct?

![reconstruction](https://user-images.githubusercontent.com/15076450/224687556-0335e5ed-e0a6-4a2f-9ef4-71aa8b735ea7.png)
