### 2023-03-15 b3766f7
Change slice naming - bug fix for reading slice in the correct order for all operating systems
