### 2023-03-15 da499d8
Change slice naming - bug fix for reading slice in the correct order for all operating systems
