### 2023-03-16 quin-med-harvard-edu/pyGRASP#4 @cariyurek
Hi @kerkelae ,

This looks good! But not on sagittal and transverse planes... Yes, there is definitely a problem with the slice order here. Although slice ordering is correct when I run the same reconstruction code. So what I suspect is that it is reading the slices in the order of s-0.mat s-1.mat s-10.mat s-11.mat ... instead of s-1.mat s-2.mat s-3.mat s-4.mat ... in os.listdir in [slice_to_vol.py](https://github.com/quin-med-harvard-edu/pyGRASP/blob/main/slice_to_vol.py) when it is writing the 4D volume. We have always run pyGRASP only in linux based systems and we did not have this problem. Apparently, reading order changes in os.listdir with different operating systems. To overcome this bug, we added "0" for the single digit slices, so from now on it will be s-00.mat s-01.mat s-02.mat ... s-09.mat s-10.mat ..., so that it will read in the correct order for all systems. We have uploaded the fixed version of the code. We are glad that there is someone testing our software. We are very sorry for this inconvenience and we will ask you to try it once more with the updated version.
