### 2023-03-17 #4865 @kerkelae
Based on the difficulties I have experienced trying to install and run the software, I do not think that the software meets the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) in its current form. An implementation of the reconstruction algorithm could be valuable to imaging researchers, but the repository appears to be a collection of scripts that have yet to be exhaustively tested. The code should be [packaged](https://packaging.python.org/en/latest/tutorials/packaging-projects/), documented, and tested better before I can recommend its publication.
