### 2023-03-17 #4865 @kurugol
@kerkelae Thank you for your efforts to review this code but this is a code for reconstructing 4D MRI data for computation of an extremely important clinical marker of kidney function, i.e. glomerular filtration rate. This kind of computation of very large 4D MRI data (dynamic 3D MRI) requires a high performance computer due to large memory requirements and GPU or CPU parallelization. I do not think this should be even intended to run on a 16GB windows machine, it is never intended for that kind of machine. Also we have a docker version that we included in the readme file and it would be very easy to test it on a powerful machine using the docker without requiring installation of any python packages. Currently docker is the easiest way to test software on different platforms. I think it would be super quick and easy way to test this in a high performance machine or over a cloud with enough memory and also with  GPU or multiple CPU resources. Currently we run this code daily for clinical processing on our high performance linux servers without any issues. Many different users downloded and ran it without any issues. We were not planning to adapt it for a 16GB windows machine because despite the toy data we provided you, the clinical data will not fit into 16GB windows memory. If you would like to test this on a high performance machine we would be happy to help you. This code has been executed by multiple researchers on a daily basis on linux platforms without problem and has so far been used to process thousands of DCE-MRI scans.
