### 2023-03-17 quin-med-harvard-edu/pyGRASP#4 @kerkelae
I am also using Linux (Ubuntu 22.04 with Anaconda and the packages defined by you), so maybe the issue lies somewhere else. I can try to run the example next week with the updated code.
