### 2023-03-17 #4865 @kerkelae
I agree that the work is valuable and that researchers could benefit from sharing the code. Since the documentation did not mention the hardware requirements, I tried the software on a Linux computer with 32GB of memory (although I do have access to a high-performance computing cluster as well). Hardware and software requirements should be clearly communicated in the documentation.
