### 2023-03-17 #4865 @kurugol
@kerkelae I agree with you that it was our mistake that we did not mention about the hardware requirements in the documentation file assuming someone with the knowledge of the field would know this. It was certainly our mistake and we are sorry about that. When you first attempted to run it it became clear that it failed because of memory limitation but we tried to accommodate your request to run it on a limited memory machine because we were trying to be kind.  If you are willing, we would appreciate it if you can test on a linux machine possibly with >128GB memory and either >20 cores or with a state of the art GPU. Multiple cores are not required but it significantly reduces the reconstruction time when parallelized. We would be happy to guide you through this.
