### 2023-03-20 fc250fe
Adding text about how much memory is required for both datasets and expected computation time for to readme.md.
