### 2023-03-27 quin-med-harvard-edu/pyGRASP#2 @curtcorum
Just a comment:

Is it possible to have a much reduced dataset for a demo?

11.33GB or even 4.55GB is (on the order of) the full sizes used...

To verify the code is works should take a much reduced dataset, and such large datasets represent a barrier to trying the code, especially if behind a low bandwidth link.
