### 2023-03-27 #4865 @emdupre
👋 Hi @cariyurek --  I wanted to get back to you about your submission following our editorial review.

In general, the reviewers have consistently noted difficulties in installing and running the software. While this revealed communication gaps in the scope of the submission (i.e., the necessary hardware and software requirements), it has also highlighted [the organization of the code](https://github.com/openjournals/joss-reviews/issues/4865#issuecomment-1369378685) as out of sync with current [Python packaging](https://joss.readthedocs.io/en/latest/review_criteria.html?highlight=packaging#installation-instructions) guidelines.

These problems have jointly suggested the need for a deeper revision to meet the [JOSS standards of research software](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). Keeping submissions open pending major updates is discouraged at JOSS, however, because of the challenge of keeping reviewers 'on the hook'. As such, with regret, I'm afraid we need to reject the current submission.

Thank you to @kerkelae and @curtcorum for your work in getting this review started; I'm sorry we couldn't finish this process more successfully.
