### 2023-03-28 #4865 @kurugol
Dear @empudre, Thanks for your efforts in reviewing this work. Unfortunately, besides the difficulties of memory requirement due to large data size of real data, there is no other problem remaining currently. And the reviewers could not run it because of the large size of the data and computational requirements.  If the reviewers can give us their reviews, i.e. the list of problems remaining, we will work on  fixing them in our new submission. A new postdoc started in our lab and he could read the instructions and run the pyGRASP code without any problems  and assistance. In the new submission we will use a very small toy data to make it easier to run. Please provide us a list of problems remaining to improve in the new submission. Thanks.
